+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Interpolators
     - Function `trajectoryRotate()` for rotating the trajectory points along the line. Used by selectors.
     - Function `trajectoryClosestIndex()` to receive index of the closest point.
+    - Functions `trajectoryFarthest()` and `trajectoryFarthestIndex()` to receive the most distant points.
 - Penalizers
     - _Centerline_
         - Methods `sum` and `avg` for computing the resulting penalty.